fix(my-copilot): address code review findings
- Fix client component fs import by extracting news-types.ts
- Fix sitemap: filter to article-type only, remove non-existent /nyheter
- Fix /statistikk/json returning redirect instead of 401 (add to API paths)
- Add empty state to NewsFeed when filter returns no results
- Only show filter chips for categories that have content
- Fix garpegenitiv in personvern/tilgjengelighet pages (språkvask)
- Fix product name consistency (Oh-My-Nav)
