+++ apps/my-copilot/src/lib/news.ts
@@ -2,7 +2,10 @@ import fs from "fs";
 import path from "path";
 import matter from "gray-matter";
 
-export type NewsCategory = "copilot" | "nav" | "nav-pilot" | "praksis" | "oppsummering";
+export type { NewsCategory, NewsItem } from "./news-types";
+export { CATEGORY_CONFIG } from "./news-types";
+
+import type { NewsCategory, NewsItem } from "./news-types";
 
 const VALID_CATEGORIES: Set<string> = new Set<string>(["copilot", "nav", "nav-pilot", "praksis", "oppsummering"]);
 
@@ -16,19 +19,6 @@ function parseCategory(value: unknown, slug: string): NewsCategory {
     console.warn(`Unknown news category "${value}" in ${slug}.md, falling back to "copilot"`);
   }
   return "copilot";
-}
-
-export interface NewsItem {
-  slug: string;
-  title: string;
-  date: string;
-  draft: boolean;
-  category: NewsCategory;
-  excerpt: string;
-  tags: string[];
-  type: "link" | "article";
-  url?: string;
-  content?: string;
 }
 
 function resolveArticlesDir(): string {
@@ -109,14 +99,3 @@ export function getArticleSlugs(): string[] {
     })
     .filter((s): s is string => s !== null);
 }
-
-export const CATEGORY_CONFIG: Record<
-  NewsCategory,
-  { label: string; variant: "info" | "success" | "warning" | "neutral" }
-> = {
-  copilot: { label: "Copilot", variant: "info" },
-  nav: { label: "Nav", variant: "success" },
-  "nav-pilot": { label: "Nav-pilot", variant: "info" },
-  praksis: { label: "Praksis", variant: "warning" },
-  oppsummering: { label: "Oppsummering", variant: "neutral" },
-};
